refactor: use codecs across metadata providers
