+++ oxiad/coordinator/controller/shard_controller_test.go
@@ -118,7 +118,7 @@ func TestShardController(t *testing.T) {
 	s2 := &proto.DataServerIdentity{Public: "s2:9091", Internal: "s2:8191"}
 	s3 := &proto.DataServerIdentity{Public: "s3:9091", Internal: "s3:8191"}
 
-	metadata := newTestMetadata(t, memory.NewProvider[*proto.ClusterStatus](), &proto.ClusterConfiguration{})
+	metadata := newTestMetadata(t, memory.NewProvider(provider.ClusterStatusCodec), &proto.ClusterConfiguration{})
 
 	sc := NewShardController(constant.DefaultNamespace, shard, namespaceConfig, &proto.ShardMetadata{
 		Status:   proto.ShardStatusUnknown,
@@ -196,7 +196,7 @@ func TestShardController_StartingWithLeaderAlreadyPresent(t *testing.T) {
 	n2 := rpc.GetNode(s2)
 	n3 := rpc.GetNode(s3)
 
-	metadata := newTestMetadata(t, memory.NewProvider[*proto.ClusterStatus](), &proto.ClusterConfiguration{})
+	metadata := newTestMetadata(t, memory.NewProvider(provider.ClusterStatusCodec), &proto.ClusterConfiguration{})
 
 	sc := NewShardController(constant.DefaultNamespace, shard, namespaceConfig, &proto.ShardMetadata{
 		Status:   proto.ShardStatusSteadyState,
@@ -227,7 +227,7 @@ func TestShardController_NewTermWithNonRespondingServer(t *testing.T) {
 	s2 := &proto.DataServerIdentity{Public: "s2:9091", Internal: "s2:8191"}
 	s3 := &proto.DataServerIdentity{Public: "s3:9091", Internal: "s3:8191"}
 
-	metadata := newTestMetadata(t, memory.NewProvider[*proto.ClusterStatus](), &proto.ClusterConfiguration{})
+	metadata := newTestMetadata(t, memory.NewProvider(provider.ClusterStatusCodec), &proto.ClusterConfiguration{})
 
 	sc := NewShardController(constant.DefaultNamespace, shard, namespaceConfig, &proto.ShardMetadata{
 		Status:   proto.ShardStatusUnknown,
@@ -274,7 +274,7 @@ func TestShardController_NewTermFollowerUntilItRecovers(t *testing.T) {
 	s2 := &proto.DataServerIdentity{Public: "s2:9091", Internal: "s2:8191"}
 	s3 := &proto.DataServerIdentity{Public: "s3:9091", Internal: "s3:8191"}
 
-	metadata := newTestMetadata(t, memory.NewProvider[*proto.ClusterStatus](), &proto.ClusterConfiguration{})
+	metadata := newTestMetadata(t, memory.NewProvider(provider.ClusterStatusCodec), &proto.ClusterConfiguration{})
 
 	sc := NewShardController(constant.DefaultNamespace, shard, namespaceConfig, &proto.ShardMetadata{
 		Status:   proto.ShardStatusUnknown,
@@ -329,7 +329,7 @@ func TestShardController_VerifyFollowersWereAllFenced(t *testing.T) {
 	n2 := rpc.GetNode(s2)
 	n3 := rpc.GetNode(s3)
 
-	metadata := newTestMetadata(t, memory.NewProvider[*proto.ClusterStatus](), &proto.ClusterConfiguration{})
+	metadata := newTestMetadata(t, memory.NewProvider(provider.ClusterStatusCodec), &proto.ClusterConfiguration{})
 
 	sc := NewShardController(constant.DefaultNamespace, shard, namespaceConfig, &proto.ShardMetadata{
 		Status:   proto.ShardStatusSteadyState,
@@ -366,7 +366,7 @@ func TestShardController_NotificationsDisabled(t *testing.T) {
 	s3 := &proto.DataServerIdentity{Public: "s3:9091", Internal: "s3:8191"}
 
 	notificationsEnabled := false
-	metadata := newTestMetadata(t, memory.NewProvider[*proto.ClusterStatus](), &proto.ClusterConfiguration{
+	metadata := newTestMetadata(t, memory.NewProvider(provider.ClusterStatusCodec), &proto.ClusterConfiguration{
 		Namespaces: []*proto.Namespace{
 			{
 				Name:                 "default",
@@ -408,7 +408,7 @@ func TestShardController_SwapNodeWithLeaderElectionFailure(t *testing.T) {
 	s3 := &proto.DataServerIdentity{Public: "s3:9091", Internal: "s3:8191"}
 	s4 := &proto.DataServerIdentity{Public: "s4:9091", Internal: "s4:8191"}
 
-	metadata := newTestMetadata(t, memory.NewProvider[*proto.ClusterStatus](), &proto.ClusterConfiguration{})
+	metadata := newTestMetadata(t, memory.NewProvider(provider.ClusterStatusCodec), &proto.ClusterConfiguration{})
 
 	sc := NewShardController(constant.DefaultNamespace, shard, namespaceConfig, &proto.ShardMetadata{
 		Status:   proto.ShardStatusUnknown,
@@ -506,7 +506,7 @@ func TestShardController_LeaderElectionShouldNotFailIfRemoveFails(t *testing.T) 
 	s3 := &proto.DataServerIdentity{Public: "s3:9091", Internal: "s3:8191"}
 	s4 := &proto.DataServerIdentity{Public: "s4:9091", Internal: "s4:8191"}
 
-	metadata := newTestMetadata(t, memory.NewProvider[*proto.ClusterStatus](), &proto.ClusterConfiguration{})
+	metadata := newTestMetadata(t, memory.NewProvider(provider.ClusterStatusCodec), &proto.ClusterConfiguration{})
 
 	sc := NewShardController(constant.DefaultNamespace, shard, namespaceConfig, &proto.ShardMetadata{
 		Status:   proto.ShardStatusUnknown,
@@ -626,7 +626,7 @@ func TestShardController_ShardsDataLostWithChangeEnsemble(t *testing.T) {
 	s5 := &proto.DataServerIdentity{Public: "s5:9091", Internal: "s5:8191"}
 	s6 := &proto.DataServerIdentity{Public: "s6:9091", Internal: "s6:8191"}
 
-	metadata := newTestMetadata(t, memory.NewProvider[*proto.ClusterStatus](), &proto.ClusterConfiguration{})
+	metadata := newTestMetadata(t, memory.NewProvider(provider.ClusterStatusCodec), &proto.ClusterConfiguration{})
 	sc := NewShardController(constant.DefaultNamespace, shardId, namespaceConfig, &proto.ShardMetadata{
 		Status:   proto.ShardStatusUnknown,
 		Term:     1,
@@ -724,7 +724,7 @@ func TestShardController_FeatureNegotiation_AllNodesSupport(t *testing.T) {
 	rpc.GetNode(s2).SetNodeFeatures(feature.SupportedFeatures())
 	rpc.GetNode(s3).SetNodeFeatures(feature.SupportedFeatures())
 
-	metadata := newTestMetadata(t, memory.NewProvider[*proto.ClusterStatus](), &proto.ClusterConfiguration{})
+	metadata := newTestMetadata(t, memory.NewProvider(provider.ClusterStatusCodec), &proto.ClusterConfiguration{})
 
 	// Create a feature supplier that queries the mock RPC
 	featureSupplier := func(servers []*proto.DataServerIdentity) map[string][]proto.Feature {
@@ -780,7 +780,7 @@ func TestShardController_FeatureNegotiation_MixedVersions(t *testing.T) {
 	// s3 is an old node without feature support
 	rpc.GetNode(s3).SetNodeFeatures([]proto.Feature{})
 
-	metadata := newTestMetadata(t, memory.NewProvider[*proto.ClusterStatus](), &proto.ClusterConfiguration{})
+	metadata := newTestMetadata(t, memory.NewProvider(provider.ClusterStatusCodec), &proto.ClusterConfiguration{})
 
 	featureSupplier := func(servers []*proto.DataServerIdentity) map[string][]proto.Feature {
 		result := make(map[string][]proto.Feature)
